backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
supabase==2.3.0
psycopg2-binary==2.9.9
pgvector==0.2.4
groq==0.4.0
google-generativeai==0.3.2
huggingface-hub==0.20.0
sentence-transformers==2.2.2
numpy==1.26.0
python-docx==1.1.0
PyPDF2==3.0.1
pypdf==4.0.0
python-magic==0.4.27
pydub==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.26.0
tenacity==8.2.3
structlog==24.1.0
qrcode==7.4.2
pillow==10.2.0
